Add initial implementation of webserver subpackage
This package is supposed to contain all the classes related to
requesting information from the webserver. This includes sending
requests and processing the returned JSON objects into Java objects.

It also defines two exceptions:
1. ResponseException
2. SendRequestException
The first exception is raised if the returned response is invalid (i.e.
null or 404). The second is raised if an error occurs during the
requesting of a resource on the webserver (i.e. no connection).